requirements.txt
alembic==1.4.2
Babel==2.8.0
beautifulsoup4==4.9.1
blinker==1.4
cachelib==0.1.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dominate==2.5.1
fake-useragent==0.1.11
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
google==3.0.0
google-api-core==1.21.0
google-auth==1.19.2
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-cloud-vision==1.0.0
google-resumable-media==0.5.1
google-search-results-serpwow==1.1.11
googleapis-common-protos==1.52.0
idna==2.10
imagetyperzapi3==0.5
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
passlib==1.7.2
protobuf==3.12.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
speaklater==1.3
SQLAlchemy==1.3.18
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
